% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRN_analysis_FigR.r
\name{plot_dorc_genes}
\alias{plot_dorc_genes}
\title{Plot DORC and RNA expression for genes}
\usage{
plot_dorc_genes(results, genes = NULL, umap_reduction = "UMAP.WNN")
}
\arguments{
\item{results}{Output from run_figr_analysis()}

\item{genes}{Character vector of genes to plot (default: top 15 valid DORCs)}

\item{umap_reduction}{Name of UMAP reduction in ATAC_se (default: "UMAP.WNN")}
}
\value{
List of combined ggplot objects
}
\description{
Plot DORC and RNA expression for genes
}
