% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnalysisToolkit-package.R
\docType{package}
\name{scAnalysisToolkit-package}
\alias{scAnalysisToolkit}
\alias{scAnalysisToolkit-package}
\title{scAnalysisToolkit: Comprehensive Toolkit for Single-Cell Multi-Omics Analysis}
\description{
For internal use by the DeCarvalho nnd Constantinides labs at Scripps Research. Provides a comprehensive suite of helper functions for single-cell RNA-seq and ATAC-seq analysis workflows. Includes helper tools/wrappers for clustering, cell type annotation with SingleR, differential expression analysis (pseudobulk, Wilcoxon, bulk RNA-seq), differential accessibility analysis, pathway enrichment (GO, Metascape), T-cell receptor repertoire analysis, cell-cell communication analysis with CellChat, motif enrichment, cNMF gene program analysis, and advanced visualizations including volcano plots, scatterplots, heatmaps, and circos diagrams. Designed to streamline end-to-end single-cell analysis with consistent interfaces and automated output generation, enabling iterative frictionless analysis of datasets. It builds on on functionality provided by single cell analysis packages such as Seurat, Signac, and others such as SingleR, DESeq2, CellChat. It operates with Seurat objects in R only (no support for scran or scanpy).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EduardAnsaldo/scAnalysisToolkit}
  \item Report bugs at \url{https://github.com/EduardAnsaldo/scAnalysisToolkit/issues}
}

}
\author{
\strong{Maintainer}: Eduard Ansaldo \email{EAnsaldo@scripps.edu}

}
\keyword{internal}
